backend/requirements.txt
Flask==2.2.5
web3==6.8.0
python-dotenv==1.0.0
